+++ app/P2PWallet/GUI/Widgets/Aftermarket/Buyer/BidHistory.hs
@@ -611,6 +611,12 @@ bidInputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -702,6 +708,12 @@ bidInputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -916,6 +928,12 @@ bidOutputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -1004,6 +1022,12 @@ bidOutputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10