+++ app/P2PWallet/GUI/Widgets/Aftermarket/Common.hs
@@ -1491,6 +1491,12 @@ currentBidsWidget sellerAddr inspectEvt AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -1617,6 +1623,12 @@ currentBidsWidget sellerAddr inspectEvt AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10