+++ app/P2PWallet/GUI/Widgets/Aftermarket/Seller/TransactionHistory.hs
@@ -566,6 +566,12 @@ aftermarketInputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -657,6 +663,12 @@ aftermarketInputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -996,6 +1008,12 @@ aftermarketOutputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10
@@ -1084,6 +1102,12 @@ aftermarketOutputs AppModel{..} = do
             , spacer_ [width 2]
             , vstack_ [childSpacing_ 3] $ for (groupInto 4 prices) $ 
                 \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+            , filler
+            , label "Deposit:"
+                `styleBasic` [textSize 8, textColor lightGray]
+            , spacer_ [width 2]
+            , label (display $ Lovelace bidDeposit)
+                `styleBasic` [textSize 8, textColor lightGray]
             ]
         ] `styleBasic` 
             [ padding 10