+++ app/P2PWallet/GUI/Widgets/TxBuilder/AftermarketBuilder/BidUpdates.hs
@@ -101,6 +101,14 @@ bidUpdatesList reverseTickerMap = map utxoRow
                 , spacer_ [width 2]
                 , vstack_ [childSpacing_ 3] $ for (groupInto 4 bid) $ 
                     \p -> hstack_ [childSpacing_ 3] $ map priceWidget p
+                , widgetIf (not isSpotBid) $ hstack
+                    [ filler
+                    , label "Deposit:"
+                        `styleBasic` [textSize 8, textColor lightGray]
+                    , spacer_ [width 2]
+                    , label (display deposit)
+                        `styleBasic` [textSize 8, textColor lightGray]
+                    ]
                 ]
             ] `styleBasic` 
                 [ padding 10
@@ -423,6 +431,18 @@ editClaimBidWidget model@AppModel{..} = do
             `styleFocus` [border 1 customBlue]
         , spacer_ [width 3]
         , helpButton claimExpirationMsg
+        ]
+    , spacer
+    , hstack
+        [ label "Bid Deposit:"
+            `styleBasic` [textSize 10]
+        , spacer
+        , textField_ (toLensVL $ maybeLens' % _2 % #deposit)
+            [placeholder "10.0"]
+            `styleBasic` [width 200, textSize 10, bgColor customGray1]
+            `styleFocus` [border 1 customBlue]
+        , spacer_ [width 3]
+        , helpButton bidDepositMsg
         ]
     , spacer
     , hstack